# [FileCheck] Don't use regex to find prefixes