@jayfoad 2023-11-14
Nice! `test/CodeGen/AMDGPU/memory-legalizer-*` may benefit from this.