@nikic 2023-11-14
> Nice! `test/CodeGen/AMDGPU/memory-legalizer-*` may benefit from this.

I checked, and it doesn't. These tests spend most time in `llc` rather than `FileCheck`. More specifically, they spend most of the time in `-verify-machineinstrs`. `verifyUseLists()` and `getMinimalPhysRegClassLLT()` together take up about 2/3 of the profile.